+++ src/SqlVirtualMachine/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+
+## Version 2.0.0
 * Converted Az.SqlVirtualMachine to autorest-based module.
 
 ## Version 1.1.1